+++ test/sbb_status_test.dart
@@ -6,59 +6,26 @@ import 'test_app.dart';
 
 void main() {
   testWidgets('status test', (WidgetTester tester) async {
+    const String text = 'Lorem ipsum sit dolor';
     final widget = Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        SBBStatus.valid(),
-        SBBStatus.warning(),
-        SBBStatus.invalid(),
-        SBBStatus.inProgress(),
-        SBBStatus.inactive(),
-        SBBStatus.valid(
-          showIcon: false,
-          text: 'Success',
-        ),
-        SBBStatus.warning(
-          showIcon: false,
-          text: 'Warning',
-        ),
-        SBBStatus.invalid(
-          showIcon: false,
-          text: 'Failure',
-        ),
-        SBBStatus.inProgress(
-          showIcon: false,
-          text: 'In Progress',
-        ),
-        SBBStatus.inactive(
-          showIcon: false,
-          text: 'Offline',
-        ),
-        SBBStatus.valid(
-          text: 'Everything is valid',
-        ),
-        SBBStatus.warning(
-          text: 'There is a potential problem',
-        ),
-        SBBStatus.invalid(
-          text: 'Somethign needs to be corrected',
-        ),
-        SBBStatus.inProgress(
-          text: 'Process is in progress',
-        ),
-        SBBStatus.inactive(
-          text: 'Application is offline',
-        ),
+        SBBStatus.alert(text: text),
+        const SizedBox(height: sbbDefaultSpacing),
+        SBBStatus.warning(text: text),
+        const SizedBox(height: sbbDefaultSpacing),
+        SBBStatus.success(text: text),
+        const SizedBox(height: sbbDefaultSpacing),
+        SBBStatus.information(text: text),
+        const SizedBox(height: sbbDefaultSpacing),
       ],
     );
 
     await Specs.run(
-      Specs.webSpecs,
+      Specs.mobileSpecs,
       widget,
       tester,
-      'status_test',
+      'status',
       find.byType(Column).first,
-      hostType: HostPlatform.web,
     );
   });
 }